Created parameter frame that can load setting/parameter. Created setting frame. JSONStructure now calls self.generate_object after loading JSON from file. Cleared JSONCreator.py file
