+++ JSONCreator.py
@@ -3,14 +3,12 @@ from source.window.main import MainWindow
 from source.frame.main import Main
 from source.controller import Controller
 from source.json_structure import JSONStructure
-from source.frame.gui import Parameter
 
 
 if __name__ == "__main__":
     app = MainWindow()
 
-    # view = Main(app)
-    view = Parameter()
+    view = Main(app)
     model = JSONStructure()
     controller = Controller(view, model)
